requirements.txt
flask==2.0.1
flask-cors==3.0.10
rasa==3.0.0
sentence-transformers==2.2.2
transformers==4.20.1
pdfplumber==0.7.1
faiss-cpu==1.7.2
torch==1.9.0
cachetools==5.0.0
requests==2.26.0